pyproject.toml
[project]
name = "ava-webhook"
version = "0.1.0"
description = "Automated job scouting and application webhook system."
readme = "README.md"
authors = [
    { name = "brennang28", email = "bgerspac@law.cardozo.yu.edu" }
]
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "python-docx>=1.1.0",
    "python-dotenv>=1.0.1",
    "google-api-python-client>=2.125.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "python-jobspy>=1.1.58",
    "langchain-core>=0.1.42",
    "langchain-ollama>=0.1.0",
    "langgraph>=0.0.32",
    "playwright>=1.42.0",
    "pydantic>=2.6.4",
    "requests>=2.31.0",
    "pandas>=2.2.1",
]

[dependency-groups]
dev = [
    "pytest>=8.1.1",
    "ruff>=0.3.4",
]

[build-system]
requires = ["uv_build>=0.10.11,<0.11.0"]
build-backend = "uv_build"

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
